Fixed bugs

+++ Weather/Models/WeatherModel.cs
@@ -13,10 +13,10 @@ namespace Weather.Models
         public string description { get; set; }
         public double visibility { get; set; }
         public string icon { get; set; }
-        public double temp { get; set; }
-        public double feels_like { get; set; }
-        public double temp_min { get; set; }
-        public double temp_max { get; set; }
+        public int temp { get; set; }
+        public int feels_like { get; set; }
+        public int temp_min { get; set; }
+        public int temp_max { get; set; }
         public int pressure { get; set; }
         public int humidity { get; set; }
         public double WindSpeed { get; set; }
